splice endpoints to different files
