+++ links/endpoint_bridge.go
@@ -1,0 +1,51 @@
+package links
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/containernetworking/plugins/pkg/ns"
+	"github.com/vishvananda/netlink"
+)
+
+type EndpointBridge struct {
+	EndpointGeneric
+}
+
+func (e *EndpointBridge) Verify() error {
+	errs := []error{}
+	err := CheckEndpointUniqueness(e)
+	if err != nil {
+		errs = append(errs, err)
+	}
+	err = CheckBridgeExists(e.GetNode())
+	if err != nil {
+		errs = append(errs, err)
+	}
+	err = CheckEndpointDoesNotExistYet(e)
+	if err != nil {
+		errs = append(errs, err)
+	}
+	if len(errs) > 0 {
+		return errors.Join(errs...)
+	}
+	return nil
+}
+
+// CheckBridgeExists verifies that the given bridge is present in the
+// network namespace referenced via the provided nspath handle.
+func CheckBridgeExists(n Node) error {
+	return n.ExecFunction(func(_ ns.NetNS) error {
+		br, err := netlink.LinkByName(n.GetShortName())
+		_, notfound := err.(netlink.LinkNotFoundError)
+		switch {
+		case notfound:
+			return fmt.Errorf("bridge %q referenced in topology but does not exist", n.GetShortName())
+		case err != nil:
+			return err
+		case br.Type() != "bridge":
+			return fmt.Errorf("interface %s found. expected type \"bridge\", actual is %q", n.GetShortName(), br.Type())
+		}
+		return nil
+	})
+}
